made changes so that totalPrice can be shown as double

+++ src/main/java/org/example/bokningssystem/dtos/DetailedBokningDto.java
@@ -18,7 +18,7 @@ public class DetailedBokningDto {
     private LocalDate startDate;
     private LocalDate endDate;
     private long nights;
-    private int totalPrice;
+    private double totalPrice;
     private KundDto kund;
     private RumDto room;
 }
